Move StorePackageFileInBackupLocationAsync to NuGetGallery.Core (#5549)
Progress on https://github.com/NuGet/Engineering/issues/1192
